moved footer around to make for better organization
