+++ b_investor_web/src/app/footer/page.module.css
@@ -10,7 +10,6 @@
 .footer>* {
     flex: 1 1 0;
 }
-
 .boxFooter {
     border: 1px solid var(--primary-dark);
     padding: 10px;
